src/main/java/com/pavelrybakov/warehouse/domains/Company.java
package com.pavelrybakov.warehouse.domains;

import java.util.ArrayList;
import java.util.List;

public class Company {

    private int id;
    private String name;
    private List<House> houses;

    public Company() {
    }

    public Company(int id, String name) {
        this.id = id;
        this.name = name;
    }

    public Company(String name) {
        this.name = name;
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public List<House> getHouses() {
        return houses;
    }

    public void setHouses(List<House> houses) {
        this.houses = houses;
    }
}
